fix review comments

+++ pairwise_distance.h
@@ -8,6 +8,6 @@ using namespace std;
 #ifndef NCFINAL_PAIRWISE_DISTANCE_H
 #define NCFINAL_PAIRWISE_DISTANCE_H
 
-void pairwise_distance(vector<vector<int>> X, vector<vector<int>> Y);
+void pairwise_distance(vector<vector<double>> X, vector<vector<double>> Y);
 
 #endif //NCFINAL_PAIRWISE_DISTANCE_H
